Create GHBranchConfig: a gh-domain specific branch config
It has been frustrating working with BranchConfig, as it lives halfway
between being a git-specific config and a gh specific config. This commit
attempts to remedy that, boiling the GHBranchConfig down to what we really
care about in gh: the MergeRef and the PushRef.

The MergeRef determines where the branch pulls from when `git pull` is
invoked. The PushRef determines where the branch pushes to when `git push`
is invoked.

In practice, these two refs are a combination of repo and branch specific
config options. Because gh is context aware, we can determine these
successfully even when git cannot. Thus, parseGHBranchConfig builds the
two refs from the available information from the various git config
values and from the branch that gh is aware of.

Why we care is that PRs for a branch are created from a branch's PushRef:
i.e. where the changes are living upstream. This is typically an easy
thing to resolve with `git rev-parse --abbrev-ref <branch>@{push}`.
However, there are several use-cases solved for here in which this is not
sufficient, revolving around triangular workflows.

Additionally, the merge and push refs in GHBranchConfig are broken into
two parts: the Remote and the Branch. This is because the Remote can
either be given as an name (commonly "origin" or "upstream") or a URL
string. Downstream consumers of this will need to handle the name or URL
differently to resolve GitHub operations, so the struct is separated for
ease of access to this Remote component.